Improve callback and error handling
We should avoid throwing errors whenever possible, since it places a
burden on callers to add try-catch in addition to callback or Promise
error handling to cover these sorts of errors.  For some errors this is
unavoidable, but should be statically known wherever Promise and
callback are not dynamic.  This way all errors are handled uniformly.

Use nodeify (like bluebird's Promise.prototype.nodeify or When.js'
when.bindCallback) to attach callback to the Promise rather than calling
it from inside the Promise.  This is necessary to avoid causing
unhandledRejection when the caller doesn't use the Promise, to avoid
callback affecting the returned Promise value, and so errors thrown by
the callback cause uncaughtException as they would normally.

Add more tests to cover callback/Promise behavior and rework argument
error tests to expect asynchronous errors.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>