Use bluebird Promise for unhandledRejection tests
The native Promise type didn't cause unhandledRejection until
nodejs/node@872702d9 (v1.4.1).  This caused the tests to fail on Node
versions which had a native Promise type but didn't emit this event
(e.g. Node v0.12).  To avoid this inconsistency in the tests, use
bluebird whenever unhandledRejection is tested for.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>