Standardize option and Promise handling
Adopt the convention that I have been using for my other modules where
options are combined using `extend` and the global Promise is used.

Although I can envision cases where the caller may want to specify a
custom Promise type, I don't expect they will be common enough to
warrant the maintenance of such an option.  If users disagree, they can
open an issue and I'll re-add it.

Remove the nodeify module.  It could be replaced with promise-nodeify,
but since Promises are not used internally, it's easier to recurse.
Note that the behavior has changed and a Promise is no longer returned
when a callback is provided.  This matches the behavior of most nodeify
functions and avoids splitting the control flow on return as well as
ambiguous error/rejection handling detection.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>