Refactor doCompare and done
Merge doneIfIncremental and doCompare into a function which takes the
comparison function and types as arguments.  This combines similar
codepaths and will make it easier to add additional ones (such as
checkpoint).

Move data/error emitting code out of done so that only cleanup and the
final 'end' is emitted from done.  This way falsey error values can be
emitted and the event emitting is closer to where the values are
created.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>