Improve and specify handling of falsey errors
If the compare function throws a falsey value, that value should be
emitted by the 'error' event, but it should not be passed as-is to the
streamCompare callback, since callers are likely to test for errors
using the truthyness of the error argument.  Instead, convert it to a
generic Error as done by promise-nodeify (since this is as good a
convention as any).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>